+++ geo-types/CHANGES.md
@@ -1,10 +1,11 @@
 # Changes
 
 ## Unreleased
-
 * `geo_types::LineString::num_coords` has been deprecated in favor of `geo::algorithm::coords_iter::CoordsIter::coords_count`
   * <https://github.com/georust/geo/pull/563>
 * Introduce `use-rstar` feature rather than `rstar` so that `approx` dependency can be optional
+  * <https://github.com/georust/geo/pull/567>
+* Implement `approx::{RelativeEq, AbsDiffEq}` for geo-types when using the `approx` feature
   * <https://github.com/georust/geo/pull/567>
 
 ## 0.6.2